A mixture distribution Keras layer, with independent logistic components.

A mixture distribution Keras layer, with independent logistic components.

layer_mixture_logistic(object, num_components, event_shape = list(),
  convert_to_tensor_fn = tfp$distributions$Distribution$sample,
  validate_args = FALSE, ...)

Arguments

object
Model or layer object

num_components
Number of component distributions in the mixture distribution.

event_shape
integer vector Tensor representing the shape of single draw from this distribution.

convert_to_tensor_fn
A callable that takes a tfd$Distribution instance and returns a tf$Tensor-like object. Default value: tfd$distributions$Distribution$sample.

validate_args
Logical, default FALSE. When TRUE distribution parameters are checked for validity despite possibly degrading runtime performance. When FALSE invalid inputs may silently render incorrect outputs. Default value: FALSE. @param ... Additional arguments passed to args of keras::create_layer.

...
Additional arguments passed to args of keras::create_layer.

See also