Seasonal state space model with effects constrained to sum to zero.

Seasonal state space model with effects constrained to sum to zero.

sts_constrained_seasonal_state_space_model(num_timesteps, num_seasons,
  drift_scale, initial_state_prior, observation_noise_scale = 1e-04,
  num_steps_per_season = 1, initial_step = 0, validate_args = FALSE,
  allow_nan_stats = TRUE, name = NULL)

Arguments

num_timesteps
Scalar integer tensor number of timesteps to model with this distribution.

num_seasons
Scalar integer number of seasons.

drift_scale
Scalar (any additional dimensions are treated as batch dimensions) float tensor indicating the standard deviation of the change in effect between consecutive occurrences of a given season. This is assumed to be the same for all seasons.

initial_state_prior
instance of tfd_multivariate_normal representing the prior distribution on latent states; must have event shape [num_seasons].

observation_noise_scale
Scalar (any additional dimensions are treated as batch dimensions) float tensor indicating the standard deviation of the observation noise.

num_steps_per_season
integer number of steps in each season. This may be either a scalar (shape []), in which case all seasons have the same length, or an array of shape [num_seasons], in which seasons have different length, but remain constant around different cycles, or an array of shape [num_cycles, num_seasons], in which num_steps_per_season for each season also varies in different cycle (e.g., a 4 years cycle with leap day). Default value: 1.

initial_step
Optional scalar integer tensor specifying the starting timestep. Default value: 0.

validate_args
logical. Whether to validate input with asserts. If validate_args is FALSE, and the inputs are invalid, correct behavior is not guaranteed. Default value: FALSE.

allow_nan_stats
logical. If FALSE, raise an exception if a statistic (e.g. mean/mode/etc...) is undefined for any batch member. If TRUE, batch members with valid parameters leading to undefined statistics will return NaN for this statistic. Default value: TRUE.

name
string prefixed to ops created by this class. Default value: "SeasonalStateSpaceModel".

Value

an instance of LinearGaussianStateSpaceModel.

See also