Decompose an observed time series into contributions from each component.

This method decomposes a time series according to the posterior represention of a structural time series model. In particular, it:

Computes the posterior marginal mean and covariances over the additive model's latent space.

Decomposes the latent posterior into the marginal blocks for each model component.

Maps the per-component latent posteriors back through each component's observation model, to generate the time series modeled by that component.

sts_decompose_by_component(observed_time_series, model, parameter_samples)

Arguments

observed_time_series
float tensor of shape concat([sample_shape, model.batch_shape, [num_timesteps, 1]]) where sample_shape corresponds to i.i.d. observations, and the trailing [1] dimension may (optionally) be omitted if num_timesteps > 1. May optionally be an instance of sts_masked_time_series, which includes a mask tensor to specify timesteps with missing observations.

model
An instance of sts_sum representing a structural time series model.

parameter_samples
list of tensors representing posterior samples of model parameters, with shapes list(tf$concat(list(list(num_posterior_draws), param<1>$prior$batch_shape, param<1>$prior$event_shape), list(list(num_posterior_draws), param<2>$prior$batch_shape, param<2>$prior$event_shape), ... ) ) for all model parameters. This may optionally also be a named list mapping parameter names to tensor values.

Value

component_dists A named list mapping component StructuralTimeSeries instances (elements of model$components) to Distribution instances representing the posterior marginal distributions on the process modeled by each component. Each distribution has batch shape matching that of posterior_means/posterior_covs, and event shape of list(num_timesteps).

See also

Examples

# \donttest{
observed_time_series <- array(rnorm(2 * 1 * 12), dim = c(2, 1, 12))
day_of_week <- observed_time_series %>% sts_seasonal(num_seasons = 7, name = "seasonal")

#> Error in py_call_impl(callable, dots$args, dots$keywords): InternalError: CUDA runtime implicit initialization on GPU:0 failed. Status: out of memory
#> 
#> Detailed traceback: 
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/sts/seasonal.py", line 831, in __init__
#>     sts_util.empirical_statistics(observed_time_series)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/sts/internal/util.py", line 221, in empirical_statistics
#>     ] = canonicalize_observed_time_series_with_mask(observed_time_series)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/sts/internal/util.py", line 315, in canonicalize_observed_time_series_with_mask
#>     observed_time_series = tf.convert_to_tensor(value=observed_time_series,
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/ops.py", line 1377, in convert_to_tensor_v2
#>     return convert_to_tensor(
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/ops.py", line 1499, in convert_to_tensor
#>     ret = conversion_func(value, dtype=dtype, name=name, as_ref=as_ref)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/tensor_conversion_registry.py", line 52, in _default_conversion_function
#>     return constant_op.constant(value, dtype, name=name)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 263, in constant
#>     return _constant_impl(value, dtype, shape, name, verify_shape=False,
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 275, in _constant_impl
#>     return _constant_eager_impl(ctx, value, dtype, shape, verify_shape)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 300, in _constant_eager_impl
#>     t = convert_to_eager_tensor(value, ctx, dtype)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 97, in convert_to_eager_tensor
#>     ctx.ensure_initialized()
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/eager/context.py", line 539, in ensure_initialized
#>     context_handle = pywrap_tfe.TFE_NewContext(opts)

local_linear_trend <- observed_time_series %>% sts_local_linear_trend(name = "local_linear")

#> Error in py_call_impl(callable, dots$args, dots$keywords): InternalError: CUDA runtime implicit initialization on GPU:0 failed. Status: out of memory
#> 
#> Detailed traceback: 
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/sts/local_linear_trend.py", line 285, in __init__
#>     sts_util.empirical_statistics(observed_time_series)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/sts/internal/util.py", line 221, in empirical_statistics
#>     ] = canonicalize_observed_time_series_with_mask(observed_time_series)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/sts/internal/util.py", line 315, in canonicalize_observed_time_series_with_mask
#>     observed_time_series = tf.convert_to_tensor(value=observed_time_series,
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/ops.py", line 1377, in convert_to_tensor_v2
#>     return convert_to_tensor(
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/ops.py", line 1499, in convert_to_tensor
#>     ret = conversion_func(value, dtype=dtype, name=name, as_ref=as_ref)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/tensor_conversion_registry.py", line 52, in _default_conversion_function
#>     return constant_op.constant(value, dtype, name=name)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 263, in constant
#>     return _constant_impl(value, dtype, shape, name, verify_shape=False,
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 275, in _constant_impl
#>     return _constant_eager_impl(ctx, value, dtype, shape, verify_shape)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 300, in _constant_eager_impl
#>     t = convert_to_eager_tensor(value, ctx, dtype)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 97, in convert_to_eager_tensor
#>     ctx.ensure_initialized()
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/eager/context.py", line 539, in ensure_initialized
#>     context_handle = pywrap_tfe.TFE_NewContext(opts)

model <- observed_time_series %>%
  sts_sum(components = list(day_of_week, local_linear_trend))

#> Error in sts_sum(., components = list(day_of_week, local_linear_trend)): object 'day_of_week' not found

states_and_results <- observed_time_series %>%
  sts_fit_with_hmc(
    model,
    num_results = 10,
    num_warmup_steps = 5,
    num_variational_steps = 15
    )

#> Error in sts_fit_with_hmc(., model, num_results = 10, num_warmup_steps = 5,     num_variational_steps = 15): object 'model' not found

samples <- states_and_results[[1]]

#> Error in eval(expr, envir, enclos): object 'states_and_results' not found

component_dists <- observed_time_series %>%
 sts_decompose_by_component(model = model, parameter_samples = samples)

#> Error in py_resolve_dots(list(...)): object 'model' not found

# }