ComputesY = g(X) = DCT(X), where DCT type is indicated by the type arg

The discrete cosine transform efficiently applies a unitary DCT operator. This can be useful for mixing and decorrelating across the innermost event dimension. The inverse X = g^{-1}(Y) = IDCT(Y), where IDCT is DCT-III for type==2. This bijector can be interleaved with Affine bijectors to build a cascade of structured efficient linear layers as in Moczulski et al., 2016. Note that the operator applied is orthonormal (i.e. norm='ortho').

tfb_discrete_cosine_transform(validate_args = FALSE, dct_type = 2,
  name = "dct")

Arguments

validate_args
Logical, default FALSE. Whether to validate input with asserts. If validate_args is FALSE, and the inputs are invalid, correct behavior is not guaranteed.

dct_type
integer, the DCT type performed by the forward transformation. Currently, only 2 and 3 are supported.

name
name prefixed to Ops created by this class.

Value

a bijector instance.

References

Moczulski M, Denil M, Appleyard J, de Freitas N. ACDC: A structured efficient linear layer. In International Conference on Learning Representations, 2016.

See also