ComputesY = g(X) = NormalCDF(x)

This bijector maps inputs from [-inf, inf] to [0, 1]. The inverse of the bijector applied to a uniform random variable X ~ U(0, 1) gives back a random variable with the Normal distribution:

tfb_normal_cdf(validate_args = FALSE, name = "normal")

Arguments

validate_args
Logical, default FALSE. Whether to validate input with asserts. If validate_args is FALSE, and the inputs are invalid, correct behavior is not guaranteed.

name
name prefixed to Ops created by this class.

Value

a bijector instance.

Details

Y ~ Normal(0, 1) pdf(y; 0., 1.) = 1 / sqrt(2 * pi) * exp(-y ** 2 / 2)

See also