Chi distribution

The Chi distribution is defined over nonnegative real numbers and uses a degrees of freedom ("df") parameter.

tfd_chi(df, validate_args = FALSE, allow_nan_stats = TRUE,
  name = "Chi")

Arguments

df
Floating point tensor, the degrees of freedom of the distribution(s). df must contain only positive values.

validate_args
Logical, default FALSE. When TRUE distribution parameters are checked for validity despite possibly degrading runtime performance. When FALSE invalid inputs may silently render incorrect outputs. Default value: FALSE.

allow_nan_stats
Logical, default TRUE. When TRUE, statistics (e.g., mean, mode, variance) use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if one or more of the statistic's batch members are undefined.

name
name prefixed to Ops created by this class.

Details

Mathematical Details

The probability density function (pdf) is,

pdf(x; df, x >= 0) = x**(df - 1) exp(-0.5 x**2) / Z
Z = 2**(0.5 df - 1) Gamma(0.5 df)

where:

df denotes the degrees of freedom,

Z is the normalization constant, and,

Gamma is the gamma function.

The Chi distribution is a transformation of the Chi2 distribution; it is the distribution of the positive square root of a variable obeying a Chi distribution.

See also