Covariance.

Covariance is (possibly) defined only for non-scalar-event distributions. For example, for a length-k, vector-valued distribution, it is calculated as, Cov[i, j] = Covariance(X_i, X_j) = E[(X_i - E[X_i]) (X_j - E[X_j])] where Cov is a (batch of) k x k matrix, 0 <= (i, j) < k, and E denotes expectation.

tfd_covariance(distribution, ...)

Arguments

distribution
The distribution being used.

...
Additional parameters passed to Python.

Value

Floating-point Tensor with shape [B1, ..., Bn, k, k] where the first n dimensions are batch coordinates and k = reduce_prod(self.event_shape).

Details

Alternatively, for non-vector, multivariate distributions (e.g., matrix-valued, Wishart), Covariance shall return a (batch of) matrices under some vectorization of the events, i.e., Cov[i, j] = Covariance(Vec(X)_i, Vec(X)_j) = [as above] where Cov is a (batch of) k x k matrices, 0 <= (i, j) < k = reduce_prod(event_shape), and Vec is some function mapping indices of this distribution's event dimensions to indices of a length-k vector.

See also

Examples

# \donttest{
d <- tfd_normal(loc = c(1, 2), scale = c(1, 0.5))

#> Error in py_call_impl(callable, dots$args, dots$keywords): InternalError: CUDA runtime implicit initialization on GPU:0 failed. Status: out of memory
#> 
#> Detailed traceback: 
#>   File "<decorator-gen-184>", line 2, in __init__
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/distributions/distribution.py", line 334, in wrapped_init
#>     default_init(self_, *args, **kwargs)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/distributions/normal.py", line 138, in __init__
#>     self._loc = tensor_util.convert_nonref_to_tensor(
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow_probability/python/internal/tensor_util.py", line 114, in convert_nonref_to_tensor
#>     return tf.convert_to_tensor(
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/ops.py", line 1377, in convert_to_tensor_v2
#>     return convert_to_tensor(
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/ops.py", line 1499, in convert_to_tensor
#>     ret = conversion_func(value, dtype=dtype, name=name, as_ref=as_ref)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 338, in _constant_tensor_conversion_function
#>     return constant(v, dtype=dtype, name=name)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 263, in constant
#>     return _constant_impl(value, dtype, shape, name, verify_shape=False,
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 275, in _constant_impl
#>     return _constant_eager_impl(ctx, value, dtype, shape, verify_shape)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 300, in _constant_eager_impl
#>     t = convert_to_eager_tensor(value, ctx, dtype)
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/framework/constant_op.py", line 97, in convert_to_eager_tensor
#>     ctx.ensure_initialized()
#>   File "/home/key/.virtualenvs/tf2.3/lib64/python3.8/site-packages/tensorflow/python/eager/context.py", line 539, in ensure_initialized
#>     context_handle = pywrap_tfe.TFE_NewContext(opts)

d %>% tfd_variance()

#> Error in eval(lhs, parent, parent): object 'd' not found

# }