Distribution representing the quantization Y = ceiling(X)

Definition in Terms of Sampling

tfd_quantized(distribution, low = NULL, high = NULL,
  validate_args = FALSE, name = "QuantizedDistribution")

Arguments

distribution
The base distribution class to transform. Typically an instance of Distribution.

low
Tensor with same dtype as this distribution and shape able to be added to samples. Should be a whole number. Default NULL. If provided, base distribution's prob should be defined at low.

high
Tensor with same dtype as this distribution and shape able to be added to samples. Should be a whole number. Default NULL. If provided, base distribution's prob should be defined at high - 1. high must be strictly greater than low.

validate_args
Logical, default FALSE. When TRUE distribution parameters are checked for validity despite possibly degrading runtime performance. When FALSE invalid inputs may silently render incorrect outputs. Default value: FALSE.

name
name prefixed to Ops created by this class.

Details

1. Draw X
2. Set Y <-- ceiling(X)
3. If Y < low, reset Y <-- low
4. If Y > high, reset Y <-- high
5. Return Y

Definition in Terms of the Probability Mass Function

Given scalar random variable X, we define a discrete random variable Y supported on the integers as follows:

 P[Y = j] := P[X <= low],  if j == low,
          := P[X > high - 1],  j == high,
          := 0, if j < low or j > high,
          := P[j - 1 < X <= j],  all other j.

Conceptually, without cutoffs, the quantization process partitions the real line R into half open intervals, and identifies an integer j with the right endpoints:

 R = ... (-2, -1](-1, 0](0, 1](1, 2](2, 3](3, 4] ...
 j = ...      -1      0     1     2     3     4  ...

P[Y = j] is the mass of X within the jth interval. If low = 0, and high = 2, then the intervals are redrawn and j is re-assigned:

 R = (-infty, 0](0, 1](1, infty)
 j =          0     1     2

P[Y = j] is still the mass of X within the jth interval.

@section References:

Tim Salimans, Andrej Karpathy, Xi Chen, and Diederik P. Kingma. PixelCNN++: Improving the PixelCNN with discretized logistic mixture likelihood and other modifications. International Conference on Learning Representations_, 2017.

Aaron van den Oord et al. Parallel WaveNet: Fast High-Fidelity Speech Synthesis. arXiv preprint arXiv:1711.10433, 2017.

See also