Survival function.

Given random variable X, the survival function is defined: tfd_survival_function(x) = P[X > x] = 1 - P[X <= x] = 1 - cdf(x).

tfd_survival_function(distribution, value, name = "survival_function")

Arguments

distribution
The distribution being used.

value
float or double Tensor.

name
String prepended to names of ops created by this function.

Value

a Tensor of shape sample_shape(x) + self$batch_shape with values of type self$dtype.

See also