A Transformed Distribution.

A TransformedDistribution models p(y) given a base distribution p(x), and a deterministic, invertible, differentiable transform,Y = g(X). The transform is typically an instance of the Bijector class and the base distribution is typically an instance of the Distribution class.

tfd_transformed(distribution, bijector, batch_shape = NULL,
  event_shape = NULL, validate_args = FALSE, name = NULL)

Arguments

distribution
The base distribution instance to transform. Typically an instance of Distribution.

bijector
The object responsible for calculating the transformation. Typically an instance of Bijector.

batch_shape
integer vector Tensor which overrides distribution batch_shape; valid only if distribution.is_scalar_batch().

event_shape
integer vector Tensor which overrides distribution event_shape; valid only if distribution.is_scalar_event().

validate_args
Logical, default FALSE. When TRUE distribution parameters are checked for validity despite possibly degrading runtime performance. When FALSE invalid inputs may silently render incorrect outputs. Default value: FALSE.

name
The name for ops managed by the distribution. Default value: bijector.name + distribution.name.

Details

A Bijector is expected to implement the following functions:

forward,

inverse,

inverse_log_det_jacobian.

The semantics of these functions are outlined in the Bijector documentation.

We now describe how a TransformedDistribution alters the input/outputs of a Distribution associated with a random variable (rv) X. Write cdf(Y=y) for an absolutely continuous cumulative distribution function of random variable Y; write the probability density function pdf(Y=y) := d^k / (dy_1,...,dy_k) cdf(Y=y) for its derivative wrt to Y evaluated at y. Assume that Y = g(X) where g is a deterministic diffeomorphism, i.e., a non-random, continuous, differentiable, and invertible function. Write the inverse of g as X = g^{-1}(Y) and (J o g)(x) for the Jacobian of g evaluated at x.

A TransformedDistribution implements the following operations:

sample Mathematically: Y = g(X) Programmatically: bijector.forward(distribution.sample(...))

log_prob Mathematically: (log o pdf)(Y=y) = (log o pdf o g^{-1})(y) + (log o abs o det o J o g^{-1})(y) Programmatically: (distribution.log_prob(bijector.inverse(y)) + bijector.inverse_log_det_jacobian(y))

log_cdf Mathematically: (log o cdf)(Y=y) = (log o cdf o g^{-1})(y) Programmatically: distribution.log_cdf(bijector.inverse(x))

and similarly for: cdf, prob, log_survival_function, survival_function.

See also